checkin for in-database pipelines
This is still very broken. Mongo code is isolated to queryMongo.js. Theoretically other database drivers can be swapped in. In-progress conversion of menu loading code to a pure function (should not read or write to global state, since global state is different between d3m and eventdata). Menu is treated as a primitive pipeline step similar to 'transform', 'subset' or 'aggregate'