mongo_retrieve_util updates and implement dataverse upload endpoints
I removed the hardcoded aggregation query inside mongo_retrieve_util, added a query reformatter for converting date and ObjectIds in extended json format to objects, and added a deserialization step for date objects returned from Mongo. The dataset list is alphabetically ordered. Implemented the api calls for uploading to dataverse in the front-end, and re-added the UI. Removed the secondary mongo_retrieve_util2. Username defaults to '', if the environment variables are not set mongo_retrieve_util attempts to load from local mongo.