deferred modelSelectionResults calls
modelSelectionResults is called on mithril redraw if 1) the results menu is active AND 2) if app.solverPending is true. The app.solverPending variable is set when a problem is selected, the current problem's pipeline is edited, or the current problem's Task, Subtask, or Metric is changed. callSolver will now use the manipulations endpoint if a nominal variable is set. Changed behavior of modelSelectionResults to wait for callSolver to resolve, instead of waiting 2 seconds.