rework how current problem is handled. Resolves #471 #448
Previously when no problem is selected, then selectedProblem is undefined. New behavior is to include the default page load as a problem listed in the discovery table, and the selectedProblem a copy of the defaultProblem. selectedProblem should only ever be undefined briefly on page load. When a problem is submitted, a copy of the currentProblem is saved to the All Problems table, and highlighted